+++ packages/eyes-webdriverio-5/src/fluent/SelectorByElement.js
@@ -1,6 +1,6 @@
 'use strict'
 
-const {GeneralUtils, GetSelector} = require('@applitools/eyes-sdk-core')
+const {GeneralUtils, GetSelector, EyesJsBrowserUtils} = require('@applitools/eyes-sdk-core')
 
 const EYES_SELECTOR_TAG = 'data-eyes-selector'
 
@@ -28,6 +28,11 @@ class SelectorByElement extends GetSelector {
       this._element.element,
     )
     return `[${EYES_SELECTOR_TAG}="${randId}"]`
+  }
+
+  async toPersistedRegions(driver) {
+    const xpath = await EyesJsBrowserUtils.getElementXpath(driver, this._element.element)
+    return [{type: 'xpath', selector: xpath}]
   }
 }