+++ packages/sdk-release-kit/src/versions/scripts/verify-versions.js
@@ -5,10 +5,9 @@ const fs = require('fs')
 const {makePackagesList, verifyDependencies} = require('..')
 const pkgs = makePackagesList()
 const results = []
-const rootFolder = process.cwd()
 
-module.exports = isFix => {
-  verifyDependencies({pkgs, pkgPath: rootFolder, results})
+module.exports = ({isFix, pkgPath}) => {
+  verifyDependencies({pkgs, pkgPath, results})
 
   const errors = results.filter(({depVersion, sourceVersion}) => depVersion !== sourceVersion)